Moved `RazeError` into an `error` module
