+++ impl/src/lib.rs
@@ -14,6 +14,7 @@
 
 pub mod bazel;
 pub mod context;
+pub mod error;
 pub mod license;
 pub mod metadata;
 pub mod planning;
