+++ impl/src/util.rs
@@ -225,74 +225,6 @@ pub fn find_bazel_workspace_root() -> Option<PathBuf> {
   return None;
 }
 
-pub const PLEASE_FILE_A_BUG: &str =
-  "Please file an issue at github.com/google/cargo-raze with details.";
-
-#[derive(Debug)]
-pub enum RazeError {
-  Generic(String),
-  Internal(String),
-  Rendering {
-    crate_name_opt: Option<String>,
-    message: String,
-  },
-  Planning {
-    dependency_name_opt: Option<String>,
-    message: String,
-  },
-  Config {
-    field_path_opt: Option<String>,
-    message: String,
-  },
-}
-
-impl std::error::Error for RazeError {}
-
-impl fmt::Display for RazeError {
-  fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-    match &self {
-      Self::Generic(s) => write!(f, "Raze failed with cause: \"{}\"", s),
-      Self::Internal(s) => write!(
-        f,
-        "Raze failed unexpectedly with cause: \"{}\". {}",
-        s, PLEASE_FILE_A_BUG
-      ),
-      Self::Config {
-        field_path_opt,
-        message,
-      } => match field_path_opt {
-        Some(path) => write!(
-          f,
-          "Raze config problem in field \"{}\" with cause: \"{}\"",
-          path, message
-        ),
-        None => write!(f, "Raze config problem with cause: \"{}\"", message),
-      },
-      Self::Rendering {
-        crate_name_opt,
-        message,
-      } => match crate_name_opt {
-        Some(name) => write!(
-          f,
-          "Raze failed to render crate \"{}\" with cause: \"{}\"",
-          name, message
-        ),
-        None => write!(f, "Raze failed to render with cause: \"{}\"", message),
-      },
-      Self::Planning {
-        dependency_name_opt,
-        message,
-      } => match dependency_name_opt {
-        Some(dep_name) => write!(
-          f,
-          "Raze failed to plan crate \"{}\" with cause: \"{}\"",
-          dep_name, message
-        ),
-        None => write!(f, "Raze failed to render with cause: \"{}\"", message),
-      },
-    }
-  }
-}
 
 pub struct PlatformDetails {
   target_triple: String,
